learning/19_oauth/requirements.txt
google-adk
google-api-python-client
google-auth-httplib2
google-auth-oauthlib 